+++ src/level/RLCommunityVotePopup.cpp
@@ -1,7 +1,6 @@
 #include "RLCommunityVotePopup.hpp"
 #include "../include/RLAchievements.hpp"
 #include "../include/RLConstants.hpp"
-#include "../layer/RLVotesLeaderboardLayer.hpp"
 #include <Geode/binding/GJAccountManager.hpp>
 #include <Geode/binding/UploadActionPopup.hpp>
 #include <algorithm>
@@ -299,15 +298,6 @@ bool RLCommunityVotePopup::init() {
     infoBtn->setPosition({25.f, 25.f});
     m_buttonMenu->addChild(infoBtn, 3);
 
-    // votes leaderboard
-    auto leaderboardSpr =
-        CCSprite::createWithSpriteFrameName("RL_lbVote01.png"_spr);
-    leaderboardSpr->setScale(0.7f);
-    auto leaderboardBtn = CCMenuItemSpriteExtra::create(
-        leaderboardSpr, this, menu_selector(RLCommunityVotePopup::onLeaderboard));
-    leaderboardBtn->setPosition({65.f, 25.f});
-    m_buttonMenu->addChild(leaderboardBtn, 3);
-
     // single toggle for moderators to show/hide all scores at once
     if (rl::isUserClassicRole() || rl::isUserPlatformerRole() || rl::isUserOwner()) {
         auto allSpr = CCSprite::createWithSpriteFrameName("hideBtn_001.png");
@@ -325,14 +315,6 @@ bool RLCommunityVotePopup::init() {
     }
 
     return true;
-}
-
-void RLCommunityVotePopup::onLeaderboard(CCObject*) {
-    auto leaderboardLayer = RLVotesLeaderboardLayer::create();
-    auto scene = CCScene::create();
-    scene->addChild(leaderboardLayer);
-    auto transitionFade = CCTransitionFade::create(0.5f, scene);
-    CCDirector::sharedDirector()->pushScene(transitionFade);
 }
 
 void RLCommunityVotePopup::refreshFromServer() {